requirements.txt
aiohttp==3.9.3
beautifulsoup4==4.12.3
lxml==5.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1git status